opa-react-demo: don't use eopa via /eopa prefix (#20)
Previously, the docker-compose setup would have /eopa reverse-proxied
to http://eopa:8181; and it worked fine. However, the dev-mode setup,
`npm start`, didn't do that, and as a result the batch demo failed if
you started `eopa run --server` yourself and used `npm start` to bring
the frontend app up.

Now, the docker-compose setup rev-proxies /v1, which is just a normal
part of the EOPA API path /v1/batch/data/...

This way, it works fine with the simple "proxy" directive of our
package.json.